diffval: Vegetation Patterns

Find, visualize and explore patterns of differential taxa in vegetation data (namely in a phytosociological table), using the Differential Value (DiffVal). Patterns are searched through mathematical optimization algorithms. Ultimately, Total Differential Value (TDV) optimization aims at obtaining classifications of vegetation data based on differential taxa, as in the traditional geobotanical approach (Monteiro-Henriques 2025, https://doi.org/10.3897/VCS.140466). The Gurobi optimizer, as well as the R package 'gurobi', can be installed from https://www.gurobi.com/products/gurobi-optimizer/. The useful vignette Gurobi Installation Guide, from package 'prioritizr', can be found here: https://prioritizr.net/articles/gurobi_installation_guide.html.

See also

Author

Maintainer: Tiago Monteiro-Henriques tmh.dev@icloud.com (ORCID)

Authors:

Jorge Orestes Cerdeira (ORCID)

Other contributors:

Fundação para a Ciência e a Tecnologia, Portugal (<https://www.fct.pt/>) [funder]